requirements.txt
pygame
gymnasium
stable-baselines3[extra]
pyyaml
moviepy
